+++ src/utils/mod.rs
@@ -21,9 +21,11 @@ mod tempdir;
 mod net;
 mod hex;
 mod parse_env;
+mod parse_time;
 
 
 pub use utils::tempdir::create_temp_dir;
 pub use utils::net::parse_forwarded_for;
 pub use utils::hex::from_hex;
 pub use utils::parse_env::parse_env;
+pub use utils::parse_time::parse_time;